[SEQ-000] Resharper format to K&R style

+++ Sequencer/Sequencer.cpp
@@ -27,8 +27,7 @@
 #include "RenderingJob.h"
 #include "StringUtils.h"
 
-Sequencer::Sequencer(QWidget* parent) : QMainWindow(parent)
-{
+Sequencer::Sequencer(QWidget* parent) : QMainWindow(parent) {
   qDebug() << "Starting " << QCoreApplication::applicationName() << " " << QCoreApplication::applicationVersion();
 
   ui.setupUi(this);
@@ -62,46 +61,37 @@ Sequencer::Sequencer(QWidget* parent) : QMainWindow(parent)
   version_label->setAlignment(Qt::AlignRight | Qt::AlignBottom);
   ui.toolBar->addWidget(version_label);
 
-  connect(theme_selector, SIGNAL(currentIndexChanged(const QString&)), this, SLOT(on_themeSelector_currentTextChanged(const QString&)));
+  connect(theme_selector, SIGNAL(currentIndexChanged(const QString&)), this,
+          SLOT(on_themeSelector_currentTextChanged(const QString&)));
 
   // hook up Sequencer::update_diagram to the RenderingThread::render_completed signal
   connect(&this->worker_thread_, &RenderingThread::render_completed, this, &Sequencer::update_diagram);
 }
 
-Sequencer::~Sequencer()
-{
+Sequencer::~Sequencer() {
   disconnect(&this->worker_thread_, &RenderingThread::render_completed, this, &Sequencer::update_diagram);
 
   this->worker_thread_.requestInterruption();
-  if (!this->worker_thread_.wait(100))
-  {
-    this->worker_thread_.terminate();
-  }
+  if (!this->worker_thread_.wait(100)) { this->worker_thread_.terminate(); }
 }
 
-void Sequencer::on_actionCreate_New_triggered()
-{
+void Sequencer::on_actionCreate_New_triggered() {
   qDebug() << "on_actionCreate_New_triggered";
-  if (dirty_check())
-  {
+  if (dirty_check()) {
     qDebug() << "creating new document";
     this->ui.textBrowser->setText(QString());
     this->ui.textBrowser->setDocumentTitle(QString());
   }
 }
 
-void Sequencer::on_actionOpen_triggered()
-{
+void Sequencer::on_actionOpen_triggered() {
   const auto file_name = QFileDialog::getOpenFileName(this, tr("Open File"), nullptr, tr("Sequencer Files (*.seq)"));
-  if (!file_name.isEmpty() && dirty_check())
-  {
+  if (!file_name.isEmpty() && dirty_check()) {
     QFile input_file(file_name);
-    if (input_file.open(QIODevice::ReadOnly))
-    {
+    if (input_file.open(QIODevice::ReadOnly)) {
       std::vector<std::string> lines;
       QTextStream in(&input_file);
-      while (!in.atEnd())
-      {
+      while (!in.atEnd()) {
         auto line = in.readLine();
         lines.push_back(line.toStdString());
       }
@@ -117,75 +107,51 @@ void Sequencer::on_actionOpen_triggered()
   }
 }
 
-void Sequencer::on_actionSave_triggered()
-{
-  do_action_save();
-}
+void Sequencer::on_actionSave_triggered() { do_action_save(); }
 
-void Sequencer::on_actionSave_As_triggered()
-{
-  do_action_save_as();
-}
+void Sequencer::on_actionSave_As_triggered() { do_action_save_as(); }
 
-void Sequencer::on_actionAdd_Title_triggered()
-{
+void Sequencer::on_actionAdd_Title_triggered() {
   bool ok;
   QString text = QInputDialog::getText(this, tr("Diagram Title"), tr("Diagram title:"), QLineEdit::Normal, "", &ok);
-  if (ok && !text.isEmpty()) {
-    replace_header_token(":title ", text.toStdString());
-  }
+  if (ok && !text.isEmpty()) { replace_header_token(":title ", text.toStdString()); }
 }
 
-void Sequencer::on_actionAdd_Author_triggered()
-{
+void Sequencer::on_actionAdd_Author_triggered() {
   bool ok;
   QString text = QInputDialog::getText(this, tr("Author Name"),
-    tr("Author name:"), QLineEdit::Normal, "", &ok);
-  if (ok && !text.isEmpty())
-  {
-    replace_header_token(":author ", text.toStdString());
-  }
+                                       tr("Author name:"), QLineEdit::Normal, "", &ok);
+  if (ok && !text.isEmpty()) { replace_header_token(":author ", text.toStdString()); }
 }
 
-void Sequencer::on_actionAdd_Date_triggered()
-{
-  replace_header_token(":date", "");
-}
+void Sequencer::on_actionAdd_Date_triggered() { replace_header_token(":date", ""); }
 
-void Sequencer::on_themeSelector_currentTextChanged(const QString& text)
-{
+void Sequencer::on_themeSelector_currentTextChanged(const QString& text) {
   replace_header_token(":theme ", text.toStdString());
 }
 
-void Sequencer::on_actionCopy_Diagram_to_Clipboard_triggered() const
-{
+void Sequencer::on_actionCopy_Diagram_to_Clipboard_triggered() const {
   const QPixmap pixmap = this->ui.label->pixmap(Qt::ReturnByValueConstant::ReturnByValue);
-  if (!pixmap.isNull())
-  {
+  if (!pixmap.isNull()) {
     QImage image(pixmap.toImage());
     QClipboard* clipboard = QApplication::clipboard();
-    if (clipboard && !image.isNull()) {
-      clipboard->setImage(image, QClipboard::Mode::Clipboard);
-    }
+    if (clipboard && !image.isNull()) { clipboard->setImage(image, QClipboard::Mode::Clipboard); }
   }
 }
 
-void Sequencer::on_actionExport_Diagram_As_triggered()
-{
-  QString file_name = QFileDialog::getSaveFileName(this, tr("Export File"), nullptr, tr("Portable Network Graphics (PNG) (*.png)"));
+void Sequencer::on_actionExport_Diagram_As_triggered() {
+  QString file_name = QFileDialog::getSaveFileName(this, tr("Export File"), nullptr,
+                                                   tr("Portable Network Graphics (PNG) (*.png)"));
   QPixmap pic = this->ui.label->pixmap(Qt::ReturnByValueConstant::ReturnByValue);
   pic.save(file_name, "PNG");
 }
 
-void Sequencer::on_actionGrammar_triggered() const
-{
+void Sequencer::on_actionGrammar_triggered() const {
   QDesktopServices::openUrl(QUrl(QString("https://github.com/rsouth/sequencer/wiki/Grammar")));
 }
 
-void Sequencer::on_actionExample_File_triggered()
-{
-  if (dirty_check())
-  {
+void Sequencer::on_actionExample_File_triggered() {
+  if (dirty_check()) {
     const auto text =
       ":theme Default\n"
       ":title Example Sequence Diagram\n"
@@ -204,33 +170,26 @@ void Sequencer::on_actionExample_File_triggered()
   }
 }
 
-void Sequencer::on_actionAbout_triggered() const
-{
+void Sequencer::on_actionAbout_triggered() const {
   AboutDialog* about_dialog = new AboutDialog();
   about_dialog->setWindowFlags(windowFlags() & ~Qt::WindowContextHelpButtonHint & ~Qt::WindowMinimizeButtonHint);
   about_dialog->setAttribute(Qt::WA_DeleteOnClose);
   about_dialog->show();
 }
 
-void Sequencer::on_textBrowser_textChanged()
-{
+void Sequencer::on_textBrowser_textChanged() {
   const auto input_text = ui.textBrowser->toPlainText().toStdString();
   const auto rendering_job = RenderingJob(this->parent(), input_text);
 
   // update theme selector
   auto lines = StringUtils::split(input_text, "\n");
-  for (const auto& line : lines)
-  {
-    if (StringUtils::starts_with(line, ":theme "))
-    {
+  for (const auto& line : lines) {
+    if (StringUtils::starts_with(line, ":theme ")) {
       std::string theme_name = StringUtils::get_token_value(line, ":theme ");
       auto found = ui.toolBar->window()->findChild<QComboBox*>("themeSelector");
-      if (found && theme_name.length() > 1)
-      {
+      if (found && theme_name.length() > 1) {
         int comboIdx = found->findText(theme_name.c_str());
-        if (comboIdx > -1) {
-          found->setCurrentIndex(comboIdx);
-        }
+        if (comboIdx > -1) { found->setCurrentIndex(comboIdx); }
       }
     }
   }
@@ -238,39 +197,26 @@ void Sequencer::on_textBrowser_textChanged()
   this->worker_thread_.render(rendering_job);
 }
 
-bool Sequencer::do_action_save()
-{
+bool Sequencer::do_action_save() {
   QString document_title = this->ui.textBrowser->documentTitle();
-  if (document_title.isEmpty())
-  {
-    return this->do_action_save_as();
-  }
-  else
-  {
-    return this->save_source_to_file(document_title.toStdString());
-  }
+  if (document_title.isEmpty()) { return this->do_action_save_as(); }
+  return this->save_source_to_file(document_title.toStdString());
 }
 
-bool Sequencer::do_action_save_as()
-{
+bool Sequencer::do_action_save_as() {
   QString file_name = QFileDialog::getSaveFileName(this, tr("Save File"), nullptr, tr("Sequencer Files (*.seq)"));
-  if (!file_name.isEmpty()) {
-    return this->save_source_to_file(file_name.toStdString());
-  }
+  if (!file_name.isEmpty()) { return this->save_source_to_file(file_name.toStdString()); }
   return !file_name.isEmpty();
 }
 
-bool Sequencer::dirty_check()
-{
-  if (this->ui.textBrowser->document()->isModified())
-  {
+bool Sequencer::dirty_check() {
+  if (this->ui.textBrowser->document()->isModified()) {
     qDebug() << "document is dirty; asking user if they want to save";
     int ret = QMessageBox::information(this, tr("Save File?"),
-      tr("The Sequencer file has been modified.\nDo you want to save your changes?"),
-      QMessageBox::Save | QMessageBox::Discard | QMessageBox::Cancel,
-      QMessageBox::Save);
-    switch (ret)
-    {
+                                       tr("The Sequencer file has been modified.\nDo you want to save your changes?"),
+                                       QMessageBox::Save | QMessageBox::Discard | QMessageBox::Cancel,
+                                       QMessageBox::Save);
+    switch (ret) {
     case QMessageBox::Save:
       qDebug() << "user selected to save before creating new";
       return do_action_save();
@@ -292,11 +238,9 @@ bool Sequencer::dirty_check()
   return true;
 }
 
-bool Sequencer::save_source_to_file(const std::string& file_name)
-{
+bool Sequencer::save_source_to_file(const std::string& file_name) {
   QFile output_file(file_name.c_str());
-  if (output_file.open(QIODevice::WriteOnly))
-  {
+  if (output_file.open(QIODevice::WriteOnly)) {
     // get source text from UI
     auto text = this->ui.textBrowser->toPlainText();
 
@@ -313,8 +257,7 @@ bool Sequencer::save_source_to_file(const std::string& file_name)
   return false;
 }
 
-void Sequencer::replace_header_token(const std::string& token, const std::string& replacement)
-{
+void Sequencer::replace_header_token(const std::string& token, const std::string& replacement) {
   // get source text from the UI
   auto text = this->ui.textBrowser->toPlainText();
   auto lines = StringUtils::split(text.toStdString(), "\n");
@@ -328,17 +271,11 @@ void Sequencer::replace_header_token(const std::string& token, const std::string
   this->ui.textBrowser->document()->setModified(true);
 }
 
-void Sequencer::closeEvent(QCloseEvent* evt)
-{
-  if (!dirty_check()) {
-    evt->ignore();
-  }
-}
+void Sequencer::closeEvent(QCloseEvent* evt) { if (!dirty_check()) { evt->ignore(); } }
 
 // slot: update_diagram
 // updates the UI with the new QPixmap holding the diagram
-void Sequencer::update_diagram(const QPixmap& img)
-{
+void Sequencer::update_diagram(const QPixmap& img) {
   ui.label->setPixmap(img);
   ui.label->setFixedSize(img.width(), img.height());
   QApplication::processEvents();
